loginHandler function now splits responsibilities into smaller functions for better readability and maintainability.
